+++ src/emhass/forecast.py
@@ -162,7 +162,7 @@ class Forecast(object):
         self.end_forecast = (self.start_forecast + self.optim_conf['delta_forecast']).replace(microsecond=0)
         self.forecast_dates = pd.date_range(start=self.start_forecast, 
                                             end=self.end_forecast-self.freq, 
-                                            freq=self.freq).round(self.freq, ambiguous='infer', nonexistent=self.freq)
+                                            freq=self.freq).round(self.freq, ambiguous='infer', nonexistent='shift_forward')
         if params is not None:
             if 'prediction_horizon' in list(self.params['passed_data'].keys()):
                 if self.params['passed_data']['prediction_horizon'] is not None:
@@ -189,7 +189,7 @@ class Forecast(object):
             freq_scrap = pd.to_timedelta(60, "minutes") # The scrapping time step is 60min
             forecast_dates_scrap = pd.date_range(start=self.start_forecast,
                                                  end=self.end_forecast-freq_scrap, 
-                                                 freq=freq_scrap).round(freq_scrap, ambiguous='infer', nonexistent=freq_scrap)
+                                                 freq=freq_scrap).round(freq_scrap, ambiguous='infer', nonexistent='shift_forward')
             # Using the clearoutside webpage
             response = get("https://clearoutside.com/forecast/"+str(round(self.lat, 2))+"/"+str(round(self.lon, 2))+"?desktop=true")
             '''import bz2 # Uncomment to save a serialized data for tests
@@ -417,19 +417,9 @@ class Forecast(object):
                 # Setting the main parameters of the PV plant
                 location = Location(latitude=self.lat, longitude=self.lon)
                 temp_params = TEMPERATURE_MODEL_PARAMETERS['sapm']['close_mount_glass_glass']
-                # cec_modules_0 = pvlib.pvsystem.retrieve_sam('CECMod')
-                # cec_modules = pvlib.pvsystem.retrieve_sam(path=self.root + '/data/CEC Modules.csv')
-                # cols_to_keep = [elem for elem in list(cec_modules_0.columns) if elem not in list(cec_modules.columns)]
-                # cec_modules = pd.concat([cec_modules, cec_modules_0[cols_to_keep]], axis=1)
-                # with bz2.BZ2File(self.root + '/data/cec_modules.pbz2', "w") as f: 
-                #     cPickle.dump(cec_modules, f)
-                # cec_inverters_0 = pvlib.pvsystem.retrieve_sam('cecinverter')
-                # cec_inverters = pvlib.pvsystem.retrieve_sam(path=self.root + '/data/CEC Inverters.csv')
-                # with bz2.BZ2File(self.root + '/data/cec_inverters.pbz2', "w") as f: 
-                #     cPickle.dump(cec_inverters, f)
-                cec_modules = bz2.BZ2File(str(pathlib.Path(self.emhass_conf['data_path'] / 'cec_modules.pbz2')), "rb")
+                cec_modules = bz2.BZ2File(self.emhass_conf['root_path'] / '/emhass/data/cec_modules.pbz2', "rb")
                 cec_modules = cPickle.load(cec_modules)
-                cec_inverters = bz2.BZ2File(str(pathlib.Path(self.emhass_conf['data_path'] / 'cec_inverters.pbz2')), "rb")
+                cec_inverters = bz2.BZ2File(self.emhass_conf['root_path'] / '/emhass/data/cec_modules.pbz2', "rb")
                 cec_inverters = cPickle.load(cec_inverters)
                 if type(self.plant_conf['module_model']) == list:
                     P_PV_forecast = pd.Series(0, index=df_weather.index)
@@ -493,7 +483,7 @@ class Forecast(object):
         end_forecast_csv = (start_forecast_csv + self.optim_conf['delta_forecast']).replace(microsecond=0)
         forecast_dates_csv = pd.date_range(start=start_forecast_csv, 
                                            end=end_forecast_csv+timedelta(days=timedelta_days)-self.freq, 
-                                           freq=self.freq).round(self.freq, ambiguous='infer', nonexistent=self.freq)
+                                           freq=self.freq).round(self.freq, ambiguous='infer', nonexistent='shift_forward')
         if self.params is not None:
             if 'prediction_horizon' in list(self.params['passed_data'].keys()):
                 if self.params['passed_data']['prediction_horizon'] is not None: