+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Added new files for updated PV modules and inverters using PVLib
 ### Fix
 - style.css auto format and adjusted table styling
+- Changed pandas datetime rounding to nonexistent='shift_forward' to help survive DST change
 
 ## 0.8.4 - 2024-03-13
 ### Improvement